refactor(web): refactor navigation components:
- Improve components names
- Separate `SidebarNavigationMenu` from `NavigationMenus`
- Re-add `Menu` component to `NavigationMenus` to keep consistency with `SidebarNavigationMenu to keep consistency with `SidebarNavigationMenu``
