--- apps/web/src/components/NavMenus.tsx
@@ -1,17 +1,0 @@
-import { NavItemComponent } from "./Menu";
-import { SidebarMenu } from "./SidebarMenu/SidebarMenu";
-import { MENU_ITEMS } from "./content";
-
-export const NavMenusSection: React.FC = () => {
-  return (
-    <div className="flex items-center justify-between">
-      <SidebarMenu className="xl:hidden" />
-
-      <div className="hidden items-center gap-4 xl:flex">
-        {MENU_ITEMS.map((item) => (
-          <NavItemComponent {...item} key={item.label} />
-        ))}
-      </div>
-    </div>
-  );
-};
